DEV: Extract time_helpers from rails_helper
